transforms: (xz-commute) add pauli propagation (#32)
Adds pauli propagation through the `xz` gadget. Includes a purpose built
runner for pattern rewrites which avoids exponential runtime of this
procedure.

I've hard coded measurement, cnot, cz, and hadamard here. In the future
we may want to find a way to encode the propagation information as part
of the gate class.