dialects: (qu) Add qu.reg type and register operations (#94)
Closes #52.

This PR introduces support for statically-sized quantum registers to the
qu dialect, as proposed in the issue.

The core of this PR is the introduction of the qu.reg type and four
fundamental operations to manipulate these registers. This allows the IR
to represent and operate on groups of qubits, aligning with other
quantum IRs like QIRO and QSSA.

---------

Co-authored-by: Alex Rice <alexrice999@hotmail.co.uk>